Examplar
Files that can be used with the program. I used a web scraper, exported data as csv file, and got rid of the columns and rows i didn't need. since i only parsed up until the first comma the most important thing is to ensure that the first column is the usernames.